refactor : LottoGenerator의 역할이 domain에 더 적합함
