+++ src/main/java/lotto/domain/Buyer.java
@@ -2,7 +2,6 @@ package lotto.domain;
 
 import lotto.wrapper.Count;
 import lotto.wrapper.Money;
-import lotto.controller.LottoGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
